extension cleanup/reorg
