+++ extensions/_d/load.inc
@@ -1,0 +1,15 @@
+<?php
+/**
+ * @file load.inc Fork/daemonize bass class.
+ * @author Stackware, LLC
+ * @version 5.0
+ * @copyright Copyright (c) 2012-2023 Stackware, LLC. All Rights Reserved.
+ * @copyright Licensed under the GNU General Public License
+ * @copyright See COPYRIGHT.txt and LICENSE.txt.
+ */
+namespace asm\promptd;
+
+define('ASM_EXT__D_ROOT',__DIR__.DIRECTORY_SEPARATOR);
+
+require(ASM_EXT__D_ROOT.'corsp.inc');
+
