+++ extensions/promptd/load.inc
@@ -1,0 +1,17 @@
+<?php
+/**
+ * @file load.inc CLI GPT with local socket/pipe and prompt templating.
+ * @author Stackware, LLC
+ * @version 5.0
+ * @copyright Copyright (c) 2012-2023 Stackware, LLC. All Rights Reserved.
+ * @copyright Licensed under the GNU General Public License
+ * @copyright See COPYRIGHT.txt and LICENSE.txt.
+ */
+namespace asm\extensions\promptd;
+
+define('ASM_EXT_PROMPTD_ROOT',__DIR__.DIRECTORY_SEPARATOR);
+
+require(ASM_EXT_PROMPTD_ROOT.'lib.php');
+require(ASM_EXT_PROMPTD_ROOT.'promptd.php');
+require(ASM_EXT_PROMPTD_ROOT.'promptdio.php');
+
