+++ extensions/promptd/promptdio.php
@@ -7,7 +7,8 @@
  * @copyright Licensed under the GNU General Public License
  * @copyright See COPYRIGHT.txt and LICENSE.txt.
  */
-namespace asm\promptd;
+namespace asm\extensions\promptd;
+
 
 
 /**
